+++ app/src/main/java/com/example/booklify/model/CategoryModel.java
@@ -4,7 +4,7 @@ public class CategoryModel {
     int author;
     int category;
     private String image;
-    double price;
+    int price;
     String title,library,content,id;
     boolean popularity;
 
@@ -24,7 +24,7 @@ public class CategoryModel {
         return title;
     }
 
-    public double getPrice() {
+    public int getPrice() {
         return price;
     }
